Fix order of checks in git diff parsing.
One of our old review requests was causing an exception where it was indexing
past the end of the diff. The issue here is that we added code to check for
new file/moved file/mode change state, but that code was added above the index
check instead of below it.

Testing done:
Loaded the review request that was causing the error. Before this change, it
caused an exception. After, it loads fine.

Reviewed at https://reviews.reviewboard.org/r/5523/