Tweak the account forms a bit.
The forms on "My Account" had a few issues. The first one I noticed was that
the label for "Keep this information private" ended with a colon, despite the
check-box being on the left side of the label. This was because we were using
django's label_tag method, which is kind of dumb and expects the form field to
always be on the right-hand side of the label. I've switched us over to using
djblets' label_tag tag, which does the right thing for check-boxes. As a nice
side-effect, we automatically get better styling for required field labels.

The other thing I noticed was some issues with alignment and sizing. The labels
were all indented 0.5em from the left-hand side of the boxes that they labeled,
which looked ugly. The e-mail field also had a different font size than the
other fields.

Testing done:
Looked at the various forms in "My Account" and saw that they all looked nice.

Reviewed at https://reviews.reviewboard.org/r/5526/