Make a bunch of updates to the user manual.
This change updates a lot of the text in our docs to match the 2.0 reality
better. This doesn't yet update any of the images--I'll go through and take a
bunch of new screenshots in a later change. There are also some larger changes
that I'd like to make to the content, which I'll be filing issues about to
track.

This includes one small change to a label on the profile page for accounts,
since when I was writing the docs, I realized that some installs may hide the
fields that are referenced by "Keep this information private".

Testing done: Built HTML docs and read through them.

Reviewed at https://reviews.reviewboard.org/r/5527/