Show the username instead of first name in the top navbar.
This mirrors the usability change we made for RBCommons a little while
ago. People know their first name, but sometimes forget the assigned
username. Instead of showing the first name in the top navbar next to
the gravatar, we now show the username instead. This particularly helps
if a first name is not set.