User manual part 2.
This change updates some of the text and replaces all of the screenshots for
the getting-started/, dashboard/, and review-requests/ directories in the user
manual. As part of this, I've pulled in the retina_images extension, which
lets us have @2x versions of our screenshots.

Testing done:
* Built HTML docs
* Checked that @2x images were included in the _images directory.

Reviewed at https://reviews.reviewboard.org/r/5541/