User manual part 3: Update "Reviewing Diffs"
This change adds a bunch of new content for the "Reviewing Diffs" page to
explain several of our coolest diff viewer features. This also rewrites some of
the text to be clearer, and replaces all of the old screenshots with both 1x
and 2x screenshots of the modern UI.

Testing done:
Built the HTML manual and looked at all the pretty things.

Reviewed at https://reviews.reviewboard.org/r/5545/