Support parallel-installed Django eggs
If a system has different versions of Django installed as separate eggs (such
as having the latest version as well as an older version for compatibility with
ReviewBoard), this addition to the WSGI loader enables it to find the matching
supported version on the system.

This will also make things easier on distro upgrades, as it will be possible to
have both the Django 1.4 egg on the system with Review Board 1.7.21 and Django
1.6 in place for Review Board 2.0

A release note will be needed for existing sites, as if they want to support
this capability, they will need to update their reviewboard.wsgi file manually.
This patch will take affect only for new installs.

Testing Done:
Installed the prototype parallel-installable Django 1.4 package on Fedora 20:
http://toshio.fedorapeople.org/django/python-django14-1.4.8-1.fc20.1.src.rpm

Verified that ReviewBoard starts properly. It failed without this patch.

Reviewed at https://reviews.reviewboard.org/r/5520/