Fix unicode and style issues in the markdown review UI.
The fix I did for markdown to properly handle unicode objects failed if the
object we were trying to iterate over was already encoded to utf-8. The
exception was caught and logged, but the resulting rendered view was empty.

I've changed the markdown UI to pass the error through to the user, so they at
least see something.

Once I got the file displaying, I noticed that there were a few issues with the
style rules. First, lists had their bullets overlapping with the line numbers
(because of the all-elements rule to get rid of padding). Second, we have a
bare 'h1' rule in common.less that made h1 smaller than h2. I've restored it to
the same size it should have been, but we should probably make the common.less
rule much more specific.

Testing done:
* Verified that without the encoding fix, I saw an error in the "Rendered" tab
  instead of nothing.
* Checked that with the encoding fix, I could see the page.
* Checked that the style changes fixed the issues I saw with this particular
  document.

Reviewed at https://reviews.reviewboard.org/r/5548/