rb-site upgrade: do not throw error if no sites are configured
Previously, if 'rb-site upgrade --all-sites' was called and the site file
either did not exist or contained no valid entries (such as all sites having
been removed but the sitelist being missed), it would return with error code 1.

This breaks scripts such as that on an upcoming Fedora release of Review Board
that will be running the --all-sites upgrade as part of the Apache start-up. In
order for it to be possible for Review Board to be installed on the system,
unconfigured and not break Apache, we need to treat this exit code as success.

Testing Done:
Verified that running 'rb-site upgrade --all-sites' with no
/etc/reviewboard/sites file printed the message "No Review Board sites listed
in /etc/reviewboard/sites" and exited with code 0 (success).

Reviewed at https://reviews.reviewboard.org/r/5547/