Show and hide the "Download Diff" link depending on the interdiff state.
When I made the diff revision selector operate without reloading the page, I
neglected the "Download Diff" link at the top of the page. This link should
only be shown when the page is displaying a full revision, and not when there's
an interdiff.

I've changed it to use an inline style to hide the link on the first page load
if appropriate, and JS to show/hide it when the displayed revision changes.

Testing done:
Played around with the revision selector and watched the link show and hide
itself. Verified that if I initially loaded an interdiff, that the link was
present in the DOM to show later.

Fixes bug 3269.

Reviewed at https://reviews.reviewboard.org/r/5560/