Fix the style for moved replace lines.
An older style from my initial move enhancement change earlier in 2.0
caused lines to have excess padding where they shouldn't.

The intent, I believe, was to add some padding between the move flag and
the text, but why this was only on replace lines, I'm not sure. This
change fixes the style by marking lines where a move flag is shown, and
adding 4x of padding between it and the text on both columns.