Fix incorrect labelling on review request status graph in admin.
The admin site has a widget that shows the percentage of all review requests in
various states. Unfortunately, this graph shows requests with a status of "D"
as drafts instead of as discarded.

I've fixed this graph to split things based on public or non-public, which
actually shows the draft state. I've then renamed the "D" state to be correct.

Testing done: Looked at the graph in the admin site using data from r.rb.o.

Fixes bug 3270.