Fix an XSS vulnerability in the search results.
Users with full names containing HTML would result in that HTML being
processed by the browser when auto-completing them in the Search field
at the top of any page. This could result in a script in the First or
Last Name fields being executed.

This change fixes the formatItem call to properly escape.

This vulnerability was made public by the reporter both publicly on the
issue tracker (+ mailing list), and on reviews.reviewboard.org. We're
going to cut a new release including this fix quickly as a result.