Allow posting multiple review requests with the same commit ID.
POSTing a new review request isn't supposed to attempt to create or
update from a commit ID unless create_from_commit_id is passed in the
request. We weren't checking this flag, though, before looking up an
existing review request with that commit ID. This triggered a
ChangeNumberInUseError, which ended up confusing RBTools.

We no longer make this check unless being told to create or update from
the commit ID.