Add some logging and better strings to the reviews models.
This ports over some of the RBCommons changes made to the reviews
models.

In DefaultReviewer, "target people" has been changed to "target users."

In Group, invite_only now has help text.

There's also some logging for access control for ReviewRequest and
Group, which can be helpful in spotting users trying to get to things
they're not allowed to get to.