Split out the headerbar and navbar into new templates.
This takes the headerbar and navbar and puts them into their own
templates. It makes it a lot easier to read the base template,
particularly with regards to the ordering of blocks and template hooks,
and it also makes it easier to modify these bars without major merge
issues.

It also fixes an issue with overriding the "navbar" block, which
would have resulted in an extra </div>, due to an ordering issue that
wasn't apparent before this change.