Handle the inability to load the siteconfig better.
My changes to automatically apply https based on the siteconfig made the
experience a bit worse when there was a problem loading said siteconfig. The
major problem was that load_siteconfig() would return silently if there was
some unexpected error. This wasn't a huge issue before, but now that we have
middleware that needs the siteconfig to load, it was showing me at
AttributeError for NoneType.

I've added some logging to load_siteconfig, and changed the middleware to pass
through if it couldn't load.

Testing done:
Tried to run my devserver against a mysql server which wasn't running. Saw the
error in the log about loading siteconfig, and saw that instead of getting a
weird AttributeError in my browser, I saw an error about the sql server.

Reviewed at https://reviews.reviewboard.org/r/5604/