Implement drag and drop on admin site (back end)
The user can now customize their dashboard by dragging and dropping widgets to
move them around.

Testing Done:
I opened Review Board on two different browsers, Chrome and Firefox, and logged
in as the same user on both browsers. On the Chrome browser, I visited the
admin dashboard and changed the ordering of the widgets with drag-and-drop. On
the Firefox browser, I also visited the admin dashboard, and I checked to make
sure that the positions of the widgets were the same on both browsers,
indicating that the positions had been saved on the back end side.

Reviewed at https://reviews.reviewboard.org/r/5528/