Skip dimmed whitespace chunks in keyboard navigation.
When "Hide whitespace changes" is selected, changes to trailing whitespace are
shown as equal instead of a replace. Unfortunately, the next/previous keyboard
shortcuts were still stopping to highlight these chunks.

This change fixes up _getNextAnchor to take into account whether or not the
parent chunk of the anchor is dimmed or not.

Testing done:
Created a diff with some whitespace changes, hid those changes, and then
browsed around with the keyboard. Made sure that it worked as expected when the
changes were shown.

Fixes bug 3280.

Reviewed at https://reviews.reviewboard.org/r/5640/