Require the extensions static directory.
The manual server update checks now look for the extensions' static
directory along with the old legacy directory, and require it to be
writable by the web server before Review Board can load.

rb-site now creates this directory on install or upgrade. We still don't
set the permissions, but instead tell the user to do this (after
install).

We're also removing the direct Pipeline dependency, since Djblets
depends on this now.