Attempt to make directories writable on rb-site upgrade.
When upgrading an existing site, there should be at least one directory
("uploaded") that is writable by the web server. We can use this as a
reference point for all other directories that need to be made writable.

We now stat that directory and change ownership of all the other
directories.

I've also added uploaded/files as a generated directory. Technically,
this can be created on-the-fly, but since we know we want it, and make
the other directories, we may as well make this one up-front.