Allow continuing when hitting a DB sync error in rb-site.
A problem people sometimes hit is forgetting to create the Review Board
database or setting the right permissions before they proceed with
rb-site install. This can lead to a failure when synchronizing the
database, and then the need to delete the site directory and start over.

We now present a helpful error with the ability to continue and try
again as many times as needed.

Fixdes bug #2126