Fix unit tests to cope with extensions static media changes.
The changes to the location for extensions static media caused a huge number of
unit tests to break, due to the updates required checks. This change creates
both the current and legacy media directories as part of the test
initialization, and updates one test that checks the number of failing
updates checks to handle the fact that there are now two media directories that
get checked.

Testing done: Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/5658/