Switch to the HTML5 DTD.
We were using the XHTML 1 Transitional DTD for the longest time for
Review Board. This ends up putting the browsers into a mode where
some of the more recent rules won't apply. For the most part, we're
unaffected by any of these quirks, but I ran into something I was
working on where this did matter. It's also more likely to impact us in
the future.