Add a check to find a review request with a matching commit ID for webhooks.
Previously, we relied only on the regex to match a review request ID in the
commit message. Now, we use the regex to try to find a matching review request
ID in the commit message, and if no match is found, we then use the commit ID
to try to find a matching review request.

Testing Done:
I pushed different commits to a GitHub repository with the post-receive webhook
enabled (and saw the expected behaviour):
- Commit that did not match the regex but matched a review request's commit ID
- Commit that did not match the regex and did not match a review request's
  commit ID
- Commit that matched the regex

Reviewed at https://reviews.reviewboard.org/r/5666/