Improve the note field for GitHub auth tokens.
When registering GitHub auth tokens, we send a note indicating what the
token is for. This was always hard-coded, making it hard to
differentiate tokens.

It also seems as if recently, GitHub started colliding tokens with the
same note.

To fix this, we now include part of the URL in the note, and also a
UUID. This should help prevent any future collisions.

Reviewed at https://reviews.reviewboard.org/r/5687/