Use siteconfig_settings for settings access.
We had hacks around the templates for falling back on defaults if
settings were missing from siteconfig. This was missing in some cases,
and resulted in breakages in the case of displaying gravatars (a recent
change).

Djblets now provides siteconfig_settings, which allows access with
falling back on defaults. This change updates all templates to use it.

Reviewed at https://reviews.reviewboard.org/r/5694/