Add a JavaScript resource for DefaultReviewer.
This adds DefaultReviewer as a resource. It doesn't cover the full
featureset of the API, but does allow for basic interaction with the
resource (mainly, basic fetching and deletion).

This isn't used today in Review Board, but may be useful down the road,
and is being used in RBCommons. I've opted to add it here, since its
size is minimal, and it helps complete our JavaScript API support, which
will only help extensions.