Improve the style of generated e-mails.
E-mails were a bit ugly. (They still are with this change, but less so.)

The border of the review request boxes was pure black, which was a bit
heavy. It was also very boxy.

I've made the colors match what we have in Review Board, and added some
border radiuses (for clients that support it) to all the main boxes.

I've also fixed up the style for the "Review request closed" bar, which
was too large, due to default styles for the <h1>.

Reviewed at https://reviews.reviewboard.org/r/5720/