Make code line numbers work as links
I added two functions to the file, one adds the links to the code line numbers
and the another one removes the link when the code line number isn't hovered.
The link is combined from the file id and the line number, which should be
enough to identify the row, which should be highlighted. The link is removed
from the th element, when user hovers on another row.

Testing Done:
Testing is done on my development server.

Reviewed at https://reviews.reviewboard.org/r/5576/