Support building media with parallel-installed Django
On systems such as Fedora where older versions of Django are installed as
alternative versions (and not directly importable), we need to ensure that we
are importing exactly the right version of the Django modules.

This patch adds a new variable in the reviewboard __init__.py that is imported
by both setup.py and build-media.py to ensure that they are both relying on the
same version of Django.

Testing Done:
Built successfully on Fedora.

Reviewed at https://reviews.reviewboard.org/r/5690/