Fix dependency checks in manage.py
Our dependency checks (which print errors and warnings when running the
devserver) weren't working, because of the changes in how the settings module
is defined. We used to check whether the settings module was defined in the
environment, but when we switched to Django 1.5, we started inserting that into
the environment ourselves. I've added a flag to check whether or not we're in a
subprocess.

Testing done:
Ran the devserver and saw it run the dependency checks (but only once).

Reviewed at https://reviews.reviewboard.org/r/5770/