Jump to and highlight anchors in the diffviewer URL hash.
When people are passing around links to specific lines in the diff viewer, it
helps to be able to see what that link is referring to. This change adds some
highlighting to the given row, and some logic to add in the anchor if it's not
present, since the templates only add anchors for lines that have comments on
them.

Testing Done:
Testing done on my development server locally, works fine there

Reviewed at https://reviews.reviewboard.org/r/5706/