Fix the post-commit creation to also create a draft.
Using the post-commit UI would create a ReviewRequest object with public=False,
but wouldn't create an associated draft. This meant that the review request
looked like it was public but wasn't.

Testing done:
Created a post-commit review request and saw the "This review request is not
yet public" banner.

Reviewed at https://reviews.reviewboard.org/r/5780/