Fix displaying and publishing drafts for initial review requests.
A review request without a draft that is marked as public=False would be
shown with a draft banner on RB 1.7, but the work on 2.0 broke that.
This was partially due to a change when reorganizing the
review_header.html template, and was further hindered by some missing
logic in the views.

Now, a review request created through the New Review Request page
without a draft will once again have a banner, and the Publish and
Discard buttons will work again.

It also fixes an issue with the error callback for Publish. It turns out
that it didn't have the correct parameters.

Reviewed at https://reviews.reviewboard.org/r/5781/