Add input escaping to the bugs section of review
If you create a new review and type a string that should be escaped into the
bugs secion (e.g. <A>), it disappears. Can be fixed by using view.formatText
that escapes input. Note that even though the text that you entered isn't
displayed, if you refresh the review, it's there, and is displayed, as the
template has "|safe" in it. The problem is only visible after immediate value
update.

There are a couple of other fields that might have the same issue, but none of
them (depends on, target groups) are likely to have input that should be
escaped in them.

Testing Done:
Before:
1) Created a new review.
2) Click edit on bug description.
3) Enter "<A>" into the field and press enter.
4) Text disappears.

After:
1) Created a new review.
2) Click edit on bug description.
3) Enter "<A>" into the field and press enter.
4) Text is there, just as expected.

Also tested behaviour with multiple inputs like "<123> <456>".

Bugs closed: 3244

Reviewed at https://reviews.reviewboard.org/r/5774/