Fix the styling of titles in the admin UI.
Titles in the admin UI were previously styled by the 'h1.title' rule,
which we replaced with '.box-title'. However, the admin UI doesn't use
boxes in the same sense as the rest of the UI, due to Django's built-in
styling.

This adds an explicit rule for titles in the admin UI, allowing them
once again to use the same styles as our other boxes.

Reviewed at https://reviews.reviewboard.org/r/5801/