Send emails to both users and mailing lists for review groups.
When we initially created review groups, we made the assumption that all groups
would have an associated mailing list. This made a lot of sense when
transitioning from an email based code review system. Shortly after, we made
it so that any review group that didn't have a mailing list would instead send
email directly to all the members. This creates a confusing situation whereby
users are expected to subscribe to mailing lists if they're present but
otherwise get emails.

This change adds an option to review groups to specify whether emails should go
only to the list, or to the list plus every subscriber. The default for this is
to keep the current behavior.

Testing done:
* Tested evolution.
* Verified that with this check-box off, the recipient list for a review
  request assigned to a group with a mailing list contained both the mailing
  list and the group members.

Reviewed at https://reviews.reviewboard.org/r/5828/