Decode author and commit messages coming from pysvn/subvertpy.
Post-commit review request creation with SVN worked fine when the commit
messages and author names were ASCII, but failed if there were unicode
characters. I've gone through the source for both libraries and verified that
this data is provided as a string, so we need to decode it to a unicode object.

Testing done: Ran unit tests.

Fixes bug 3346.

Reviewed at https://reviews.reviewboard.org/r/5835/