Always link to the current documentation URL.
We've always had to manually update the various URLs to the user manual,
and we've never been very consistent about doing that. However, this is
pretty easy to compute, so there's not much reason for us to have to
change anything.

There's now a new get_manual_url() function that's used to populate the
URL for the web UI and for rb-site. This will use a versioned manual if
it's a released build tagged 'final', or the 'dev' manual otherwise.

Reviewed at https://reviews.reviewboard.org/r/5854/