Fix stripping of XML-like syntax in comments
If a comment consisted of only XML-like syntax (such as "<test>"), it would be
treated as if the comment was completely empty and the comment would be
canceled. The issue here is that reviewCommentFormIsEmpty would strip out tags
before testing whether the comment was empty or not. Removing the call to
stripTags() fixes the bug.

Testing Done:
I manually tried to recreate the bug with different ways. By submitting
'<sim realtime="true>' or 'something <sim realtime="true>' in the comments you
can verify that the bug is no longer exists.

Bugs closed: 3298

Reviewed at https://reviews.reviewboard.org/r/5796/