Fix error with repo settings caused by hosting services with extra %s
When adding a new repository, having %s in the base URL of a non-custom bug
tracker breaks things. That's because we don't have an error checking for that,
and once the complete URL gets generated out of it, the problem is detected,
but that leaves the form in a broken state, where you cannot save the form
anymore.

We decided to address this by letting the user know of the error and tell
him/her that % character can be escaped if there is a need. The appropriate
error message is presented if this error happens.

Testing Done:
* Entering invalid URL for bugzilla, trac or redmine will result in an error
  that you must fix before submitting.
* Other things work as expected.

Bugs closed: 3033

Reviewed at https://reviews.reviewboard.org/r/5810/